significant speed bonus, broke out static combatant information out
correct: 58.4%, time_outs: 196
improvement: 16.8%
log loss: 2.413591

+++ simulator/src/sim/actions/mod.rs
@@ -13,7 +13,7 @@ pub enum ActionKind {
 #[derive(Copy, Clone, Debug)]
 pub struct Action {
     pub kind: ActionKind,
-    pub range: i16,
+    pub range: i8,
     pub target_id: CombatantId,
 }
